Внесены изменения по домашнему заданию спринта 8. Добавлены цвета из макета, градиент на картинки в ленте и светлый статус-бар
