+++ ImageFeed/AppDelegate.swift
@@ -16,6 +16,8 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
         return true
     }
+    
+    
 
     // MARK: UISceneSession Lifecycle
