+++ ImageFeed/Assets.xcassets/AppIcon.appiconset/Contents.json
@@ -1,6 +1,7 @@
 {
   "images" : [
     {
+      "filename" : "1024x1024@2x.png",
       "idiom" : "universal",
       "platform" : "ios",
       "size" : "1024x1024"
