+++ ImageFeed/Assets.xcassets/YPBackgroundAlpha20.colorset/Contents.json
