+++ ImageFeed/Assets.xcassets/likeButtonInactive.imageset/Contents.json
