+++ ImageFeed/Assets.xcassets/likeButtonInactive.imageset/No Active@2x.png
