+++ ImageFeed/Presentation/ImagesListCell.swift
@@ -1,0 +1,17 @@
+//
+//  ImagesListCell.swift
+//  ImageFeed
+//
+//  Created by Aleksey Yakushev on 01.05.2023.
+//
+
+import UIKit
+
+final class ImagesListCell: UITableViewCell {
+    
+    @IBOutlet weak var cellImage: UIImageView!
+    @IBOutlet weak var dateLabel: UILabel!
+    @IBOutlet weak var likeButton: UIButton!
+    
+    static let reuseIdentifier = "ImagesListCell"
+}
